+++ frontend/src/components/exercise/Exercise.js
@@ -2,12 +2,13 @@ import React from 'react';
 import { Draggable } from 'react-beautiful-dnd';
 
 const Exercise = ({ exercise, index, selectExercise }) => {
+  const { id, lift, printout } = exercise;
   const handleClick = () => {
-    selectExercise(exercise.id);
+    selectExercise(id);
   };
   return (
     <div>
-      <Draggable draggableId={exercise.id} index={index}>
+      <Draggable draggableId={id} index={index}>
         {(provided) => (
           <li
             ref={provided.innerRef}
@@ -15,7 +16,7 @@ const Exercise = ({ exercise, index, selectExercise }) => {
             {...provided.dragHandleProps}
             onClick={handleClick}
           >
-            {exercise.lift}: {exercise.printout}
+            {lift}: {printout}
           </li>
         )}
       </Draggable>
