+++ frontend/src/context/auth/AuthState.js
@@ -2,7 +2,7 @@ import React, { useReducer } from 'react';
 import axios from 'axios';
 import AuthContext from './authContext';
 import authReducer from './authReducer';
-import setAuthToken from '../../utils/setAuthToken';
+import setAuthToken from '../../functions/setAuthToken';
 
 const AuthState = (props) => {
   const initialState = {
