+++ frontend/src/functions/setAuthToken.js
@@ -1,9 +1,11 @@
 import axios from 'axios';
 
-export default token => {
+const setAuthtoken = (token) => {
   if (token) {
     axios.defaults.headers.common['x-auth-token'] = token;
   } else {
     delete axios.defaults.common['x-auth-token'];
   }
 };
+
+export default setAuthtoken;
